20251019_claude_skills_for_adk/app/requirements.txt
google-adk>=1.10.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
websockets>=13.0
